+++ client/recmake_dfclient_wasm.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-rm -rf dfclient-wasm
+if [ "$1" != "--keep" ]
+then
+    rm -rf dfclient-wasm
+fi
 
 emcmake cmake -DCMAKE_BUILD_TYPE=Debug -S dfclient -B dfclient-wasm || true
 emcmake cmake -DCMAKE_BUILD_TYPE=Debug -S dfclient -B dfclient-wasm
